Extracted NettyResponseHandler
